+++ app/shared/components/layout/header/header-menu-button.tsx
@@ -16,7 +16,8 @@ import {
   Flex,
   Button,
 } from "@/app/shared/components";
-import { MyActivities, SignoutButton } from "@/app/features/user/components";
+import { MyActivities } from "@/app/features/user/components";
+import { SignoutButton } from "@/app/features/auth/components";
 import { ThemeSelectButtons } from "../../theme";
 import { LoginModal } from "@/app/features/auth/components";
